Eliminate byte and short ranges and progressions from range construction functions' return type.

+++ core/builtins/native/kotlin/Primitives.kt
@@ -137,9 +137,9 @@ public class Byte private () : Number, Comparable<Byte> {
     public operator fun unaryMinus(): Int
 
      /** Creates a range from this value to the specified [other] value. */
-    public operator fun rangeTo(other: Byte): ByteRange
+    public operator fun rangeTo(other: Byte): IntRange
      /** Creates a range from this value to the specified [other] value. */
-    public operator fun rangeTo(other: Short): ShortRange
+    public operator fun rangeTo(other: Short): IntRange
      /** Creates a range from this value to the specified [other] value. */
     public operator fun rangeTo(other: Int): IntRange
      /** Creates a range from this value to the specified [other] value. */
@@ -277,9 +277,9 @@ public class Short private () : Number, Comparable<Short> {
     public operator fun unaryMinus(): Int
 
      /** Creates a range from this value to the specified [other] value. */
-    public operator fun rangeTo(other: Byte): ShortRange
+    public operator fun rangeTo(other: Byte): IntRange
      /** Creates a range from this value to the specified [other] value. */
-    public operator fun rangeTo(other: Short): ShortRange
+    public operator fun rangeTo(other: Short): IntRange
      /** Creates a range from this value to the specified [other] value. */
     public operator fun rangeTo(other: Int): IntRange
      /** Creates a range from this value to the specified [other] value. */
